backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
requests==2.31.0
httpx==0.27.0
groq==0.5.0
google-api-core==2.17.1